+++ changelog.md
@@ -2,7 +2,7 @@
 
 **Version 0.2.1**
 
-
+- Extended script to handle fine-tuning and from scratch-training 
 - Added a script only mode.
 
 - Added support for docs.